04-generate-questions/requirements.txt
langchain-ollama
python-dotenv